+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # CHANGELOG
 
+## 1.0.4
+* Update to android v2 embedding
+
 ## 1.0.3
 
 * Fix share image top view controller dismissal bug on iOS 13+
-* Update to use the new platform definitions
+* Update to use the new platforms definition
 
 ## 1.0.2+1